requirements.txt
akismet
amqp==5.0.5
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
dateutils==0.6.12
defusedxml==0.7.0rc2
dj-database-url==0.5.0
Django==2.2.18
django-appconf==1.0.4
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework-braces==0.3.4
django-reversion==3.0.9
django-templated-mail==1.1.1
django-timezone-field==4.1.1
django-user-accounts==3.0.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
email-reply-parser==0.5.12
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
lxml==4.6.2
Markdown==3.3.3
markdown2==2.4.0
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.1.0
prompt-toolkit==3.0.14
# psycopg2
pycparser==2.20
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz
requests==2.25.1
requests-oauthlib==1.3.0
sentry-sdk==0.20.3
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
wincertstore==0.2
zipp==3.4.0
